# api/openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener API
  description: A simple pluggable API to shorten URLs.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /shorten:
    post:
      summary: Shorten a long URL
      operationId: shortenUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortenRequest'
      responses:
        '200':
          description: URL successfully shortened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortenResponse'
        '400':
          description: Invalid request
        '500':
          description: Internal server error

  /{code}:
    get:
      summary: Redirect to the original URL
      operationId: redirectUrl
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          description: The short code representing the long URL
      responses:
        '302':
          description: Redirect to the original URL
          headers:
            Location:
              description: Redirect location
              schema:
                type: string
        '404':
          description: Short URL not found

  /metrics:
    get:
      summary: Get top 3 most shortened domains
      operationId: Metrics
      responses:
        '200':
          description: List of top domains with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DomainCount'

components:
  schemas:
    ShortenRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          example: https://example.com/long/path

    ShortenResponse:
      type: object
      properties:
        short_url:
          type: string
          example: http://localhost:8080/abc123

    DomainCount:
      type: object
      properties:
        domain:
          type: string
          example: youtube.com
        count:
          type: integer
          example: 3